requirements.txt
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-playwright==0.5.2
pyotp==2.9.0
pillow==10.4.0